Jpasetup

See merge request kdg-ti/integration-4/2024-2025/team-18/ebiketeam18!4

+++ src/main/java/be/kdg/integration4/domain/Workshop.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class Workshop {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer workshopId;
+    private Long workshopId;
     private String workshopName;
     private Location workshopLocation;
